chore: move to zoci
